+++ src/App.tsx
@@ -3,10 +3,10 @@ import CardInputSection from './components/cardInfoForm/CardInputSection/CardInp
 import CardNumberField from './components/cardInfoForm/CardNumberField/CardNumberField';
 import CardValidityPeriodField from './components/cardInfoForm/CardValidityPeriodField/CardValidityPeriodField';
 import CardCVCField from './components/cardInfoForm/CardCVCField/CardCVCField';
-import { useState } from 'react';
 import styled from '@emotion/styled';
 import useCardNumber from './hooks/useCardNumber';
 import useCardValidityPeriod from './hooks/useCardValidityPeriod';
+import useCardCVC from './hooks/useCardCVC';
 
 function App() {
   const { cardNumber, onChange, checkCardNumberError, isError } =
@@ -19,26 +19,8 @@ function App() {
     checkCardValidityPeriodError,
   } = useCardValidityPeriod();
 
-  const [cardCVC, setCardCVC] = useState('');
-  const [isCardCVCError, setIsCardCVCError] = useState(false);
-
-  const onChangeCVC = (e: React.ChangeEvent<HTMLInputElement>) => {
-    const { value } = e.target;
-
-    if (value.length > 3) {
-      return;
-    }
-
-    const isNotValid =
-      value.length < 3 || value.length > 3 || Number(value) < 0;
-
-    setIsCardCVCError(isNotValid);
-    setCardCVC(value);
-  };
-
-  const checkCardCVCError = () => {
-    return isCardCVCError;
-  };
+  const { cardCVC, isCardCVCError, onChangeCVC, checkCardCVCError } =
+    useCardCVC();
 
   return (
     <AppLayout>
